fix a bunch of style nits
I'm not sure if these were intentional, but there are things in here that didn't follow a consistent pattern or are not best practices (using `errors.New(fmt.Sprintf())`).

I'm not looking to assert/force any specific style. I figured I'd open this to see if you wanted it merged in. If not, please feel free to close. Likewise, if there are any adjustments you'd like made please let me know.